BUG: Configuration CMakeLists.txt was outdated.
ImageNoise was not building anymore with recent versions of ITK (>4.6?). First, it was still using the CMake variable USE_ITK_FILE instead of the more recent ITK_USE_FILE. Second, it was lookoing for GenerateCLP after ITK. This was resetting the ITK Image IO and creating a discrepancy between the IOs that ITK was configuring at first and the IO that ImageNoise was actuallly linked against. The order in which the "find_package" is done between SlicerExecutionModel and ITK has therefore been inverted.